imc_app/requirements.txt
kivy
matplotlib
kivy-garden.matplotlib
numpy
pillow